requirements.txt
aiohttp>=3.11.16
beautifulsoup4>=4.12.2
boto3==1.37.37
botocore==1.37.37
fake_useragent==1.5.1
Flask==3.0.3
flask_cors==5.0.1
matplotlib>=3.10.1
numpy==1.26.0
orjson==3.10.16
pandas==2.2.3
pymongo==4.6.1
pymysql==1.1.1
pyspark==3.5.1
python-dotenv>=1.0.1
Requests>=2.32.3
rich>=14.0.0
rocketmq==0.4.4
scikit_learn==1.4.2
# torch==2.3.0+cu121
torch_geometric==2.5.3
tqdm==4.67.1
urllib3>=2.3.0
